LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/tables/form-in-tables-xhtml-expected.txt


FAIL Computed display of form inside table in xhtml should be 'block' assert_equals: expected "block" but got "none"
FAIL Computed display of form inside thead in xhtml should be 'block' assert_equals: expected "block" but got "none"
FAIL Computed display of form inside tbody in xhtml should be 'block' assert_equals: expected "block" but got "none"
FAIL Computed display of form inside tfoot in xhtml should be 'block' assert_equals: expected "block" but got "none"
FAIL Computed display of form inside tr in xhtml should be 'block' assert_equals: expected "block" but got "none"
